homologacao - feat: Ajustes finais do backend

+++ src/main/java/br/univesp/pi/domain/model/Telefone.java
@@ -18,5 +18,5 @@ public class Telefone {
     private int ddd;
 
     @NotNull
-    private int numero;
+    private Long numero;
 }
